# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Library API",
    "description": "REST API for managing library books and members",
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "tags": [
    {
      "name": "Auth",
      "description": "GitHub OAuth authentication endpoints"
    },
    {
      "name": "Books",
      "description": "Book management endpoints"
    },
    {
      "name": "Members",
      "description": "Library member management endpoints"
    }
  ],
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/": {
      "get": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "firstName": {
                  "example": "any"
                },
                "lastName": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "phone": {
                  "example": "any"
                },
                "membershipType": {
                  "example": "any"
                },
                "joinDate": {
                  "example": "any"
                },
                "address": {
                  "example": "any"
                },
                "active": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/auth/status": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/auth/login": {
      "get": {
        "description": "",
        "responses": {
          "503": {
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/auth/github/callback": {
      "get": {
        "description": "",
        "responses": {
          "503": {
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/auth/login-failed": {
      "get": {
        "description": "",
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/auth/profile": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/auth/logout": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/books/": {
      "get": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "example": "any"
                },
                "author": {
                  "example": "any"
                },
                "isbn": {
                  "example": "any"
                },
                "genre": {
                  "example": "any"
                },
                "publishedYear": {
                  "example": "any"
                },
                "publisher": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                },
                "pageCount": {
                  "example": "any"
                },
                "language": {
                  "example": "any"
                },
                "available": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/books/{id}": {
      "get": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "example": "any"
                },
                "author": {
                  "example": "any"
                },
                "isbn": {
                  "example": "any"
                },
                "genre": {
                  "example": "any"
                },
                "publishedYear": {
                  "example": "any"
                },
                "publisher": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                },
                "pageCount": {
                  "example": "any"
                },
                "language": {
                  "example": "any"
                },
                "available": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/members/": {
      "get": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "firstName": {
                  "example": "any"
                },
                "lastName": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "phone": {
                  "example": "any"
                },
                "membershipType": {
                  "example": "any"
                },
                "joinDate": {
                  "example": "any"
                },
                "address": {
                  "example": "any"
                },
                "active": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/members/{id}": {
      "get": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "firstName": {
                  "example": "any"
                },
                "lastName": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "phone": {
                  "example": "any"
                },
                "membershipType": {
                  "example": "any"
                },
                "joinDate": {
                  "example": "any"
                },
                "address": {
                  "example": "any"
                },
                "active": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/status": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/login": {
      "get": {
        "description": "",
        "responses": {
          "503": {
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/github/callback": {
      "get": {
        "description": "",
        "responses": {
          "503": {
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/login-failed": {
      "get": {
        "description": "",
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/profile": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/logout": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/{id}": {
      "get": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "firstName": {
                  "example": "any"
                },
                "lastName": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "phone": {
                  "example": "any"
                },
                "membershipType": {
                  "example": "any"
                },
                "joinDate": {
                  "example": "any"
                },
                "address": {
                  "example": "any"
                },
                "active": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "definitions": {
    "Book": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "example": "674a1b2c3d4e5f67890abcde"
        },
        "title": {
          "type": "string",
          "example": "The Great Gatsby"
        },
        "author": {
          "type": "string",
          "example": "F. Scott Fitzgerald"
        },
        "isbn": {
          "type": "string",
          "example": "978-0-7432-7356-5"
        },
        "genre": {
          "type": "string",
          "example": "Fiction"
        },
        "publishedYear": {
          "type": "number",
          "example": 1925
        },
        "publisher": {
          "type": "string",
          "example": "Scribner"
        },
        "description": {
          "type": "string",
          "example": "A novel about the American dream set in the Jazz Age."
        },
        "pageCount": {
          "type": "number",
          "example": 180
        },
        "language": {
          "type": "string",
          "example": "English"
        },
        "available": {
          "type": "boolean",
          "example": true
        }
      }
    },
    "BookInput": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string",
          "example": "The Great Gatsby"
        },
        "author": {
          "type": "string",
          "example": "F. Scott Fitzgerald"
        },
        "isbn": {
          "type": "string",
          "example": "978-0-7432-7356-5"
        },
        "genre": {
          "type": "string",
          "example": "Fiction"
        },
        "publishedYear": {
          "type": "number",
          "example": 1925
        },
        "publisher": {
          "type": "string",
          "example": "Scribner"
        },
        "description": {
          "type": "string",
          "example": "A novel about the American dream set in the Jazz Age."
        },
        "pageCount": {
          "type": "number",
          "example": 180
        },
        "language": {
          "type": "string",
          "example": "English"
        },
        "available": {
          "type": "boolean",
          "example": true
        }
      }
    },
    "Member": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "example": "674a1b2c3d4e5f67890abcdf"
        },
        "firstName": {
          "type": "string",
          "example": "Jane"
        },
        "lastName": {
          "type": "string",
          "example": "Doe"
        },
        "email": {
          "type": "string",
          "example": "jane.doe@example.com"
        },
        "phone": {
          "type": "string",
          "example": "555-867-5309"
        },
        "membershipType": {
          "type": "string",
          "example": "premium"
        },
        "joinDate": {
          "type": "string",
          "example": "2024-01-15"
        },
        "address": {
          "type": "string",
          "example": "123 Main Street, Anytown, USA"
        },
        "active": {
          "type": "boolean",
          "example": true
        }
      }
    },
    "MemberInput": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string",
          "example": "Jane"
        },
        "lastName": {
          "type": "string",
          "example": "Doe"
        },
        "email": {
          "type": "string",
          "example": "jane.doe@example.com"
        },
        "phone": {
          "type": "string",
          "example": "555-867-5309"
        },
        "membershipType": {
          "type": "string",
          "example": "premium"
        },
        "joinDate": {
          "type": "string",
          "example": "2024-01-15"
        },
        "address": {
          "type": "string",
          "example": "123 Main Street, Anytown, USA"
        },
        "active": {
          "type": "boolean",
          "example": true
        }
      }
    },
    "AuthStatus": {
      "type": "object",
      "properties": {
        "authenticated": {
          "type": "boolean",
          "example": true
        },
        "user": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "example": "67fabc1234def56789012345"
            },
            "githubId": {
              "type": "string",
              "example": "12345678"
            },
            "username": {
              "type": "string",
              "example": "octocat"
            },
            "displayName": {
              "type": "string",
              "example": "The Octocat"
            },
            "email": {
              "type": "string",
              "example": "octocat@github.com"
            },
            "avatarUrl": {
              "type": "string",
              "example": "https://avatars.githubusercontent.com/u/583231?v=4"
            },
            "provider": {
              "type": "string",
              "example": "github"
            }
          }
        },
        "loginUrl": {
          "type": "string",
          "example": "/auth/login"
        }
      }
    },
    "AuthUser": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": "67fabc1234def56789012345"
        },
        "githubId": {
          "type": "string",
          "example": "12345678"
        },
        "username": {
          "type": "string",
          "example": "octocat"
        },
        "displayName": {
          "type": "string",
          "example": "The Octocat"
        },
        "email": {
          "type": "string",
          "example": "octocat@github.com"
        },
        "avatarUrl": {
          "type": "string",
          "example": "https://avatars.githubusercontent.com/u/583231?v=4"
        },
        "provider": {
          "type": "string",
          "example": "github"
        }
      }
    },
    "SuccessResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "Operation completed successfully"
        }
      }
    },
    "CreatedResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "Resource created successfully"
        },
        "id": {
          "type": "string",
          "example": "674a1b2c3d4e5f67890abcde"
        }
      }
    },
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string",
          "example": "Error message here"
        }
      }
    }
  }
}